+++ app/components/index.ts
@@ -4,4 +4,5 @@ export { default as ImageSelector } from './imageSelector';
 export { default as BalancePanel } from './balancePanel';
 export { default as BottomButton } from './bottomButton';
 export { default as SectionHeader } from './sectionHeader';
+export { default as SearchInput } from './searchInput';
 export { default as Tab } from './tab';