+++ app/components/searchInput/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  infoMessage: {
+    fontSize: 20,
+    color: '#9a9b9c',
+    textAlign: 'center',
+    marginTop: 20,
+  }
+});