+++ app/resources/values/consts.ts
@@ -9,4 +9,6 @@ export const DETAIL_CATEGORY_TABS: Array<Tab> = [
     id: 2,
     title: 'Detailed'
   },
-]
+];
+
+export const SEARCH_TEXT_MINIMUM_LENGTH = 2;