+++ app/screens/searchTransaction/styles.ts
@@ -8,11 +8,5 @@ export const styles = StyleSheet.create({
   screenTopText: {
     fontSize: 25,
     marginHorizontal: 20,
-  },
-  infoMessage: {
-    fontSize: 20,
-    color: '#9a9b9c',
-    textAlign: 'center',
-    marginTop: 20,
   }
 });